image-metadata-addition/requirements.txt
kafka-python==2.0.2
requests==2.27.1
Pillow==9.0.1